A manual cherry-picking of f012b0, making use of syn_engines work mostly.
From the original commit by t0m:

Aaaaaand the logic (such as it is) for syntax stuff is pretty much is
where I wanted it.

Still needs a pile of work, but like this, we can test
mangling/highlighting engines seperately to the code which causes things
to be highlighted, which is nice as you can replace highlighting strategy
and individual highlighter implementations independently.