Bug fix for merge diffs and a test.
In the case of the merge diff fix we now check if $parent has
something to be compared against before making a comparison.

The case of the t/00git_version.t fix is a little more
interesting. For reasons I don't fully understand running a program
which writes to stdout in scalar context has $! is set to 9 (aka
"Bad file descriptor"). Something to do with close-on-exec I think.