Avoid Gitalist::Git::Repository extra work
When creating an instance last_change, owner and description where getting
built, although they are meant to be lazily built.

On machines with lots of repos in a directory, the overhead of getting those
three properties would slow down all pages (even the ones that only display
objects of one repo).