Exeprimental: More performance in repository pages
Model was getting instanced more than one time per request due to interaction
between InstancePerContext and Catalyst-View-Component-SubInclude. SubInclude
was cleaning the stash per subrequest. InstancePerContext relies on the stash
to cache the created object.

I've activated a hidden option in SubRequest module to avoid whiping the stash
This option is not officially supported until I write appropiate tests for it
and it gets accepted.